+++ src/index.tsx
@@ -18,6 +18,8 @@ type IonConf = {
 */
 const isMacLike = /(Mac|iPhone|iPod|iPad)/i.test(navigator.platform)
 
+const keyArray = Object.keys as <T>(obj: T) => (keyof T)[]
+
 const isObject = function (obj: any) {
   return !!obj && typeof obj === 'object'
 }
@@ -35,7 +37,7 @@ const assign = function<T extends Dict, U extends Dict> (dest: T, src: U) {
 }
 
 /**
- * Ensure className as string
+ * Ensure to get a non-empty `className` as an string.
  */
 const classAsStr = function (klass: string | string[]) {
   if (typeof klass === 'string') {
@@ -45,9 +47,9 @@ const classAsStr = function (klass: string | string[]) {
   return klass.filter(Boolean).join(' ')
 }
 
-/*
-  Internal configuration
-*/
+/**
+ * Internal configuration object.
+ */
 const _Conf: IonConf = {
   map: {},
   def: {
@@ -84,16 +86,25 @@ export const setDefaults = function (defaults: IonIconDefs) {
     invariant(isObject(defaults), 'The defaults must be an object.')
     const defs = assign(_Conf.def, deepClone(defaults))
 
+    // check and format class names, if any
     if (defs.className) {
-      defs.className = classAsStr(defs.className)
+      defs.className = classAsStr(defs.className).trim()
     }
-    delete (defs as Dict).innerRef
-    delete (defs as Dict).name
+
+    // cleanup empty properties
+    if (!defs.className) {
+      delete defs.className
+    }
+    keyArray(defs).forEach((k) => {
+      if (defs[k] == null) {
+        delete defs[k]
+      }
+    })
   }
 }
 
 /**
- * Reset the table of named sizes.
+ * Reset the table of named sizes. You can extend this.
  */
 export const setSizes = function (sizes: Partial<IonIconSizes>) {
   invariant(isObject(sizes), 'The sizes must be an object.')
@@ -113,33 +124,42 @@ export class IonIcon extends React.PureComponent<IonIconProps> {
     innerRef: PropTypes.func,
   }
 
-  mergeProps(props: IonIconProps) {
-    const dest = deepClone(_Conf.def)
+  /**
+   * Merge the user properties with the defaults, taking care to preserve
+   * the classes and styles of both.
+   */
+  mergeDefs(props: IonIconProps) {
+    const defs = _Conf.def
+    const keys = keyArray(defs)
 
-    let classes = dest.className || ''
-    if (props.className) {
-      classes += ` ${classAsStr(props.className)}`
+    for (let i = 0; i < keys.length; i++) {
+      const k = keys[i]
+
+      switch (k) {
+        case 'className':
+          props.className = props.className
+            ? `${classAsStr(props.className)} ${defs.className}`
+            : defs.className
+          break
+
+        case 'style':
+          props.style = props.style
+            ? assign(assign({}, props.style), defs.style!)
+            : defs.style
+          break
+
+        default:
+          if (!props.hasOwnProperty(k)) {
+            props[k] = defs[k]
+          }
+      }
     }
 
-    let style = dest.style
-    if (props.style) {
-      style = assign(style || {}, props.style)
-    }
-
-    assign(dest, props)
-
-    if (classes) {
-      dest.className = classes.trim()
-    }
-    if (style) {
-      dest.style = style
-    }
-
-    return dest as IonIconProps
+    return props
   }
 
   render() {
-    const opts = this.mergeProps(this.props)
+    const opts = deepClone(this.props) as Merge<IonIconProps>
     const name = opts.name
     delete opts.name
 
@@ -151,6 +171,9 @@ export class IonIcon extends React.PureComponent<IonIconProps> {
       delete opts.innerRef
       opts.ref = innerRef as any
     }
+
+    // name & innerRef are out, merge with defaults before color & size
+    this.mergeDefs(opts)
 
     const color = opts.color
     if (color != null) {